# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "photoo.lu API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://api.photoo.lu"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "photoo_session",
        "description": "Session cookie used by the web app"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer session token used by mobile apps"
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "example": "Request validation failed"
          },
          "details": {},
          "requestId": {
            "type": "string",
            "format": "uuid",
            "description": "UUID identifier",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "additionalProperties": false
      },
...
      "ProfileLinks": {
        "type": "object",
        "properties": {
          "instagram": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "website": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "behance": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "other": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "label": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 60
                },
                "url": {
                  "type": "string",
                  "format": "uri"
                }
              },
              "required": [
                "label",
                "url"
              ],
              "additionalProperties": false
            },
            "default": []
          }
        },
        "additionalProperties": false
      },
...
      "OwnPhotographerProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID identifier",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "slug": {
            "type": "string",
            "minLength": 3,
            "maxLength": 60,
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$",
            "description": "URL-safe profile slug",
            "example": "jane-doe-photography"
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120,
            "example": "Jane Doe Photography"
          },
          "headline": {
            "type": [
              "string",
              "null"
            ],
            "minLength": 1,
            "maxLength": 200
          },
          "bio": {
            "type": "object",
            "properties": {
              "en": {
                "type": "string",
                "minLength": 1
              },
              "fr": {
                "type": "string",
                "minLength": 1
              },
              "de": {
                "type": "string",
                "minLength": 1
              },
              "pt": {
                "type": "string",
                "minLength": 1
              },
              "es": {
                "type": "string",
                "minLength": 1
              }
            },
            "description": "Text keyed by supported locale, only translated locales present",
            "example": {
              "en": "Wedding and portrait photographer in Luxembourg"
            }
          },
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "coverUrl": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "links": {
            "$ref": "#/components/schemas/ProfileLinks"
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "wedding",
                "portrait",
                "event",
                "product",
                "real-estate",
                "corporate"
              ],
              "example": "wedding"
            }
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[a-z]{2}$",
              "description": "ISO 639-1 language code",
              "example": "en"
            }
          },
          "location": {
            "$ref": "#/components/schemas/LatLng"
          },
          "serviceRadiusKm": {
            "type": [
              "number",
              "null"
            ],
            "exclusiveMinimum": 0
          },
          "city": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "countryCode": {
            "type": "string",
            "pattern": "^[A-Z]{2}$",
            "description": "ISO 3166-1 alpha-2 country code",
            "example": "LU"
          },
          "ratingAvg": {
            "type": "number",
            "minimum": 0,
            "maximum": 5
          },
          "ratingCount": {
            "type": "integer",
            "minimum": 0
          },
          "verificationStatus": {
            "type": "string",
            "enum": [
              "unverified",
              "pending",
              "verified",
              "rejected"
            ]
          },
          "isPublished": {
            "type": "boolean"
          },
          "stripeOnboardingComplete": {
            "type": "boolean"
          },
          "stripePayoutsEnabled": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "slug",
          "displayName",
          "headline",
          "bio",
          "avatarUrl",
          "coverUrl",
          "links",
          "categories",
          "languages",
          "location",
          "serviceRadiusKm",
          "city",
          "countryCode",
          "ratingAvg",
          "ratingCount",
          "verificationStatus",
          "isPublished",
          "stripeOnboardingComplete",
          "stripePayoutsEnabled"
        ],
        "additionalProperties": false
      },
      "LatLng": {
        "type": "object",
        "properties": {
          "lat": {
            "type": "number",
            "minimum": -90,
            "maximum": 90,
            "example": 49.6116
          },
          "lng": {
            "type": "number",
            "minimum": -180,
            "maximum": 180,
            "example": 6.1319
          }
        },
        "required": [
          "lat",
          "lng"
        ],
        "additionalProperties": false
      },
...
    },
...
  },
  "paths": {
...
    "/v1/me/photographer-profile": {
      "get": {
        "summary": "Get the current user's photographer profile",
        "tags": [
          "profiles"
        ],
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The current user's photographer profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnPhotographerProfile"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create the photographer profile for the current user",
        "tags": [
          "profiles"
        ],
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "categories": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "wedding",
                        "portrait",
                        "event",
                        "product",
                        "real-estate",
                        "corporate"
                      ],
                      "example": "wedding"
                    },
                    "minItems": 1
                  },
                  "languages": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[a-z]{2}$",
                      "description": "ISO 639-1 language code",
                      "example": "en"
                    },
                    "minItems": 1
                  },
                  "location": {
                    "$ref": "#/components/schemas/LatLng"
                  },
                  "city": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "countryCode": {
                    "type": "string",
                    "pattern": "^[A-Z]{2}$",
                    "description": "ISO 3166-1 alpha-2 country code",
                    "example": "LU"
                  },
                  "headline": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "bio": {
                    "type": "object",
                    "properties": {
                      "en": {
                        "type": "string",
                        "minLength": 1
                      },
                      "fr": {
                        "type": "string",
                        "minLength": 1
                      },
                      "de": {
                        "type": "string",
                        "minLength": 1
                      },
                      "pt": {
                        "type": "string",
                        "minLength": 1
                      },
                      "es": {
                        "type": "string",
                        "minLength": 1
                      }
                    },
                    "description": "Text keyed by supported locale, only translated locales present",
                    "example": {
                      "en": "Wedding and portrait photographer in Luxembourg"
                    }
                  },
                  "links": {
                    "$ref": "#/components/schemas/ProfileLinks"
                  },
                  "serviceRadiusKm": {
                    "type": [
                      "number",
                      "null"
                    ],
                    "exclusiveMinimum": 0
                  }
                },
                "required": [
                  "displayName",
                  "categories",
                  "languages",
                  "location",
                  "city",
                  "countryCode"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Photographer profile created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnPhotographerProfile"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update the current user's photographer profile",
        "tags": [
          "profiles"
        ],
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "headline": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "bio": {
                    "type": "object",
                    "properties": {
                      "en": {
                        "type": "string",
                        "minLength": 1
                      },
                      "fr": {
                        "type": "string",
                        "minLength": 1
                      },
                      "de": {
                        "type": "string",
                        "minLength": 1
                      },
                      "pt": {
                        "type": "string",
                        "minLength": 1
                      },
                      "es": {
                        "type": "string",
                        "minLength": 1
                      }
                    },
                    "description": "Text keyed by supported locale, only translated locales present",
                    "example": {
                      "en": "Wedding and portrait photographer in Luxembourg"
                    }
                  },
                  "links": {
                    "$ref": "#/components/schemas/ProfileLinks"
                  },
                  "categories": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "wedding",
                        "portrait",
                        "event",
                        "product",
                        "real-estate",
                        "corporate"
                      ],
                      "example": "wedding"
                    },
                    "minItems": 1
                  },
                  "languages": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[a-z]{2}$",
                      "description": "ISO 639-1 language code",
                      "example": "en"
                    },
                    "minItems": 1
                  },
                  "location": {
                    "$ref": "#/components/schemas/LatLng"
                  },
                  "serviceRadiusKm": {
                    "type": [
                      "number",
                      "null"
                    ],
                    "exclusiveMinimum": 0
                  },
                  "city": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "countryCode": {
                    "type": "string",
                    "pattern": "^[A-Z]{2}$",
                    "description": "ISO 3166-1 alpha-2 country code",
                    "example": "LU"
                  },
                  "avatarUploadId": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "uuid",
                    "description": "UUID identifier",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                  },
                  "coverUploadId": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "uuid",
                    "description": "UUID identifier",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnPhotographerProfile"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/photographer-profile/portfolio": {
      "get": {
        "summary": "List the current user's portfolio images, in all statuses",
        "tags": [
          "profiles"
        ],
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of portfolio images",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PortfolioImage"
                      }
                    },
                    "nextCursor": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "minLength": 1
                    }
                  },
                  "required": [
                    "items",
                    "nextCursor"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Attach an uploaded image to the current user's portfolio",
        "tags": [
          "profiles"
        ],
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "uploadId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "UUID identifier",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                  }
                },
                "required": [
                  "uploadId"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Portfolio image attached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioImage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/photographer-profile/portfolio/order": {
      "patch": {
        "summary": "Reorder the current user portfolio images",
        "tags": [
          "profiles"
        ],
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "imageIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid",
                      "description": "UUID identifier",
                      "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    "minItems": 1
                  }
                },
                "required": [
                  "imageIds"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Portfolio reordered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PortfolioImage"
                      }
                    },
                    "nextCursor": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "minLength": 1
                    }
                  },
                  "required": [
                    "items",
                    "nextCursor"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/photographer-profile/portfolio/{imageId}": {
      "delete": {
        "summary": "Delete a portfolio image",
        "tags": [
          "profiles"
        ],
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "UUID identifier",
              "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
            },
            "required": true,
            "description": "UUID identifier",
            "name": "imageId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Portfolio image deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}